initialize prisma schema
