+++ packages/lucia-auth/src/cli/scripts/initialize/database/index.ts
@@ -1,0 +1,7 @@
+import { DATABASE } from "../constant.js";
+import { initializePrisma } from "./prisma.js";
+
+export const initializeDatabase = async (databaseId: DATABASE) => {
+	if (databaseId === DATABASE.PRISMA) return initializePrisma();
+	return false;
+};
